feat: add "expr_noreflectmethod" build tag
By enabling this tag, two features are removed:
- functions in environment
- Eval() method

Fix #863.
