+++ builtin/lib.go
@@ -562,9 +562,8 @@ func get(params ...any) (out any, err error) {
 	// Methods can be defined on any type.
 	if v.NumMethod() > 0 {
 		if methodName, ok := i.(string); ok {
-			method := v.MethodByName(methodName)
-			if method.IsValid() {
-				return method.Interface(), nil
+			if m, ok := runtime.MethodByName(v, methodName); ok {
+				return m, nil
 			}
 		}
 	}
